% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MadeOn.R
\name{MadeOn}
\alias{MadeOn}
\title{Calculating Date of Conception from Date of Birth}
\usage{
MadeOn(x, format = "\%d.\%m.\%Y", get = "point", sd = 1.96)
}
\arguments{
\item{x}{a vector of birth dates}

\item{format}{a text string explicating the date format for x, defaults to "%d.%m.%Y"}

\item{get}{either "point" or "interval" for just the point estimate or the interval boundaries, using a standard deviation of 13 days, respectively}

\item{sd}{indicates the width of the uncertainty interval as multiples of the standard deviation (13 days)}
}
\value{
a vector of estimated dates of conception (for "point") or a data.frame of dates of conception and lower and upper boundaries
}
\description{
Calculating Date of Conception from Date of Birth
}
\examples{
MadeOn("31.05.1984")
MadeOn(c("24.04.1984","01.12.1983"),get="interval",sd=1)
}
\keyword{birthday}
\keyword{conception}
\keyword{naegele}

